requirements.txt
transformers
torch
pillow
numpy
pandas
tqdm
